Add HookAddonPrefix/IsAddonPrefixRegistered
IsAddonPrefixRegistered will return true if a prefix has been registered
with the given ID internal to Chomp, since using the native WoW API to
do this doesn't necessarily mean that the prefix was registered through
Chomp.

HookAddonPrefix will, for any prefix registered through Chomp, apply an
additional function to be called _after_ the original callback as a
post-hook.

The two functions can be used in tandem to check if an addon has
registered comms and, if so, elect to also receive their data for
processing.